backend/requirements.txt
fastapi==0.104.0
uvicorn==0.24.0
sqlalchemy[asyncio]==2.0.0
pydantic>=2.0.2,<3.0.0
pydantic-settings>=2.0.0
alembic==1.12.0
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
celery==5.3.0
redis==5.0.0
minio==7.2.0
httpx==0.25.0
python-dotenv==1.0.0
prometheus_fastapi_instrumentator
email-validator 